packages/ingestkit-excel/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestkit-excel"
version = "0.1.0"
description = "Excel file ingestion plugin for the ingestkit framework"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1",
    "pandas>=2.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
qdrant = ["qdrant-client>=1.7"]
ollama = ["httpx>=0.27"]
postgres = ["psycopg2-binary>=2.9"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external services)",
    "integration: Integration tests (require external services)",
]
